Extract some functions from `tasks/components.py` into `tasks/endpoint_operations.py`
