+++ fractal_server/app/routes/api/v1/task_collection.py
@@ -17,15 +17,15 @@ from .....config import get_settings
 from .....logger import close_logger
 from .....logger import set_logger
 from .....syringe import Inject
-from .....tasks.collection import create_package_dir_pip
 from .....tasks.collection import create_package_environment_pip
-from .....tasks.collection import download_package
-from .....tasks.collection import get_collection_data
 from .....tasks.collection import get_collection_log
 from .....tasks.collection import get_collection_path
 from .....tasks.collection import get_log_path
-from .....tasks.collection import inspect_package
 from .....tasks.collection import slugify_task_name
+from .....tasks.endpoint_operations import create_package_dir_pip
+from .....tasks.endpoint_operations import download_package
+from .....tasks.endpoint_operations import get_collection_data
+from .....tasks.endpoint_operations import inspect_package
 from .....tasks.utils import _TaskCollectPip
 from ....db import AsyncSession
 from ....db import DBSyncSession
